Add a sync byte to each command
Each command is now 3 instead of 2 bytes. A third byte, the "sync byte",
is added to the end of the message. It has the value 0xFF and is used to
verify that all bytes were sent properly. This fixes the problem that
if one byte was missed, all subsequent messages would be messed up due
to the first message not having enough bytes, and thus remaining in the
buffer.