Add new commands to set the speed of the pattern
The new commands are able to change the amount at which the time is
increased every time, thus controlling the speed of the patterns.

Because the time increment is a 16 bit integer, the command is split
into 2 (CMD_SPEEDHIGH and CMD_SPEEDLOW, 0x07 and 0x08), with one setting
the higher order byte and one setting the lower order byte.